+++ lib/glimmer-libui-cc-graphs_and_charts.rb
@@ -1,4 +1,5 @@
 $LOAD_PATH.unshift(File.expand_path('..', __FILE__))
 
 require 'glimmer-dsl-libui'
+require 'glimmer/view/bar_chart'
 require 'glimmer/view/line_graph'